+++ apps/api/views/branch_views.py
@@ -219,3 +219,106 @@ def create_branch_view(request):
     except Exception as e:
         message = f'Ошибка при обработке запроса {e}'
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
+
+
+@swagger_auto_schema(
+    method="put",
+    manual_parameters=[
+        openapi.Parameter(
+            name='name',
+            in_=openapi.TYPE_STRING,
+            type=openapi.TYPE_STRING,
+            required=True,
+            description='Название филиала'
+        ),
+        openapi.Parameter(
+            name='address',
+            in_=openapi.TYPE_STRING,
+            type=openapi.TYPE_STRING,
+            required=False,
+            description='Адрес'
+        ),
+        openapi.Parameter(
+            name='phone_number',
+            in_=openapi.TYPE_NUMBER,
+            type=openapi.TYPE_NUMBER,
+            required=False,
+            description='Номер телефона'
+        ),
+        openapi.Parameter(
+            name='email',
+            in_=openapi.FORMAT_EMAIL,
+            type=openapi.FORMAT_EMAIL,
+            required=False,
+            description='Почта'
+        ),
+        openapi.Parameter(
+            name='description',
+            in_=openapi.TYPE_STRING,
+            type=openapi.TYPE_STRING,
+            required=False,
+            description='Описание'
+        ),
+        openapi.Parameter(
+            name='short_description',
+            in_=openapi.TYPE_STRING,
+            type=openapi.TYPE_STRING,
+            required=False,
+            description='Короткое описание'
+        ),
+    ],
+    responses={
+        200: openapi.Response(
+            description='Запрос выполнен успешно',
+            schema=BranchSerializer
+        ),
+        400: openapi.Response(
+            description='Ошибка при запросе'
+        ),
+        401: openapi.Response(
+            description='Пустой или неправильный токен'
+        ),
+        403: openapi.Response(
+            description='Ошибка доступа'
+        ),
+        404: openapi.Response(
+            description='Branch не найден'
+        ),
+        405: openapi.Response(
+            description='Данный метод запроса запрещен'
+        ),
+        422: openapi.Response(
+            description='Отсутствует обязательное поле'
+        )
+    },
+    operation_description='Данный endpoint изменяет информацию о branch по {id}. Если владельцем branch является не'
+                          ' текущий пользователь, будет отказано в изменении. Можно изменять не все поля.',
+    operation_summary='Изменить branch'
+)
+@api_view(['PUT'])
+@permission_classes([IsAuthenticated])
+def update_branch_view(request, pk):
+    """Контроллер для обновления информации branch"""
+    user = request.user
+    data = request.data
+
+    try:
+        branch = get_branch_by_id_service(user, pk)
+        if branch:
+            serializer = BranchSerializer(branch, many=False, partial=True, data=data)
+            if serializer.is_valid():
+                serializer.save()
+            return Response(data=serializer.data, status=status.HTTP_200_OK)
+        else:
+            return Response(data={'detail': 'Это не ваш branch'}, status=status.HTTP_403_FORBIDDEN)
+
+    except ObjectDoesNotExist as e:
+        return Response(data={'detail': f'Такой branch не найден {e}'}, status=status.HTTP_404_NOT_FOUND)
+
+    except KeyError as e:
+        message = f'Ошибка при обработке запроса. Отсутствует поле {e}'
+        return Response(data={'detail': message}, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
+
+    except Exception as e:
+        message = f'Ошибка при обработке запроса {e}'
+        return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)