+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 ## [Unreleased]
 
+## [0.0.22] - 2022-11-03
+
+### Добавлено
+
+- New endpoint (branch create) [@fringer2423](https://github.com/fringer2423)
+- Added docstrings for new service
+- New endpoint (branch update)
+- Task serializer
+- Task endpoints
+- Create branchs commands for tests
+
+### Изменено
+
+- Edited Branch model fields requirements [@fringer2423](https://github.com/fringer2423)
+- Edited all views (replace 200 -> 201 status)
+- BranchSerializer fix
+- Google tasks
 
 ## [0.0.21] - 2022-11-03
 
@@ -156,7 +173,7 @@
 - Добавлено поле rate в UserSerializer
 - Компонент Prices.js (добавлена карусель) [@AlexTrubkina](https://github.com/AlexTrubkina)
 - Компоненты Prices и HomeScreen
-- CSS buttons, main, purplecircle
+- CSS buttons, main, purple circle
 
 ## [0.0.11] - 2022-10-19